integrations/openclaw-adapter/coremind_side/pyproject.toml
[project]
name = "coremind-plugin-openclaw-adapter"
version = "0.1.0"
description = "OpenClaw adapter — CoreMind side (gRPC plugin bridging OpenClaw events and actions)."
readme = "../README.md"
license = { text = "AGPL-3.0-or-later" }
authors = [{ name = "Guillaume Gagnon" }]
requires-python = ">=3.12"
keywords = ["coremind", "openclaw", "adapter", "plugin", "integration"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "coremind>=0.0.1",
    "cryptography>=43.0.0",
    "grpcio>=1.66.0",
    "jcs>=0.2.1",
    "jsonschema>=4.23.0",
    "structlog>=24.4.0",
]

[project.scripts]
coremind-plugin-openclaw-adapter = "coremind_plugin_openclaw.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coremind_plugin_openclaw"]
